move ModelParams to utils
